requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.0
future==0.18.2
greenlet==1.0.0
kombu==4.6.11
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
